Update french locale for main.cfg
Hi Deadlock989,

In collaboration with SNova and Redstylt, also members of the Community of French-speaking Factorio players, I am pleased to provide you with the new version of the French translation of the 'main.cfg' file. It mainly consists of a few translation corrections and updates following some recent modifications of the mod.

Thank you very much for this mod which is very popular with the French-speaking players  (among others, I guess). 
Very nice work !

On behalf of the French-speaking Community, 
bev